models/hunyuan/requirements.txt
dify_plugin~=0.0.1b54
tencentcloud-sdk-python~=3.0.1294
